pyproject.toml
[project]
name = "aoi-sentinel"
version = "0.1.0"
description = "Mamba-RL false-call filter for SMT AOI (Saki) — NPI online learning under escape constraint"
requires-python = ">=3.10,<3.12"
authors = [{ name = "Jin Ho Choi" }]
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=15.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.7",
    "tqdm>=4.66",
    "python-dotenv>=1.0",

    "opencv-python>=4.9",
    "Pillow>=10.2",
    "scikit-image>=0.22",
    "albumentations>=1.4",

    "lxml>=5.1",

    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",

    "gymnasium>=0.29",
]

[project.optional-dependencies]
# Training stack — Linux/WSL only because of mamba-ssm CUDA kernels.
train = [
    "torch>=2.4,<2.6",
    "torchvision>=0.19",
    "timm>=1.0.11",
    "mambavision>=1.0",
    "mamba-ssm>=2.2.2",
    "causal-conv1d>=1.4.0",
    "scikit-learn>=1.4",
    "tensorboard>=2.16",
    "wandb>=0.17",
]
threed = [
    "open3d>=0.18",
    "trimesh>=4.3",
]
rag = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "langchain>=0.2",
    "langchain-community>=0.2",
]
# Optional offline-RL fallback for periodic batch fine-tunes.
offline_rl = [
    "d3rlpy>=2.5",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "jupyterlab>=4.2",
]

[project.scripts]
aoi = "aoi_sentinel.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aoi_sentinel*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
